Add integration tests to ensure dexing is OK
Since the stubs are hand-written, need to ensure that the manually added signatures are actually compatible with the dexer and min sdk level.

This generates java code that calls into each stub using javassist to parse stub class files (reflection is not available since jvm classloader does not allow redefining java.lang classes). The generated code is packaged into a library which is then dexed into a dummy android app.