Better integration tests and support for multiple API levels
Previous integration test was not actionable - the APK would assemble even if an unsupported API call (e.g. to java.time.Instant) was included. The new test calls D8 directly and asserts that there are no desugaring warnings (sadly, that requires reflection).

Additionally, refactoring the project so that other API builds could be included more easily and including API 21 build as an example.