Ensure that serialized signatures use HashSets
Because Animal Sniffer signatures are serialized using Java
Serialization, we need to make sure that the types embedded into the
signatures are available on the classpath when Animal Sniffer is run.
Specifically, for Clazz.signatures, when the set is empty, Kotlin's
toSet method will return a kotlin.collections.EmptySet. This change
ensures that a plain HashSet is always used.

Fixes https://github.com/open-toast/gummy-bears/issues/23.