Fix core library desugaring v2 documentation (#108)
https://developer.android.com/studio/write/java11-default-support-table
> - The minimal version includes only the `java.util.function` package
and replacement classes for some Android framework concurrent
collections that have known issues on old devices.
> - The default version is effectively an upgraded version of the Java 8
API desugaring updated to Java 11. It includes the `java.time` and
`java.util.stream` packages in addition to everything in the minimal
version.
> - The nio version includes the `java.nio` package in addition to
everything in the default version.

The nio version is actually the "full" flavor.